wrote code for WorkExperienceInfo

+++ src/main/java/com/example/demo/model/WorkExperienceInfo.java
@@ -6,9 +6,9 @@ import lombok.Data;
 @Builder
 @Data
 public class WorkExperienceInfo {
-    private int id;
-    private int resumeId;
-    private int years;
+    private Long id;
+    private Long resumeId;
+    private Integer years;
     private String companyName;
     private String position;
     private String responsibilities;
